reformat folder hierarchy and create basic homepage
