+++ src/App.tsx
@@ -1,63 +1,19 @@
 import React, { useState } from "react";
-import "./App.css";
-import { Button, Form } from "react-bootstrap";
+import "./assets/App.css";
 import { Navbar } from "./Navbar";
-import { Page } from "./interfaces/page"
-import logo from './logo.svg';
-import { Footer } from "./Footer"
-
-//local storage and API Key: key should be entered in by the user and will be stored in local storage (NOT session storage)
-let keyData = "";
-const saveKeyData = "MYKEY";
-const prevKey = localStorage.getItem(saveKeyData); //so it'll look like: MYKEY: <api_key_value here> in the local storage when you inspect
-if (prevKey !== null) {
-  keyData = JSON.parse(prevKey);
-}
+import { Page } from "./interfaces/page";
+import logo from "./logo.svg";
+import { Footer } from "./Footer";
+import { Homepage } from "./Homepage";
 
 function App() {
-  const [key, setKey] = useState<string>(keyData); //for api key input
-  const [page, setPage] = useState<Page>("api");
-
-  //sets the local storage item to the api key the user inputed
-  function handleSubmit() {
-    localStorage.setItem(saveKeyData, JSON.stringify(key));
-    window.location.reload(); //when making a mistake and changing the key again, I found that I have to reload the whole site before openai refreshes what it has stores for the local storage variable
-  }
-
-  //whenever there's a change it'll store the api key in a local state called key but it won't be set in the local storage until the user clicks the submit button
-  function changeKey(event: React.ChangeEvent<HTMLInputElement>) {
-    setKey(event.target.value);
-  }
+  const [page, setpage] = useState<Page>("home");
   return (
-    <div className="App">
-    <Navbar page={page} setPage={setPage}></Navbar>
-    { page === "api" &&
-    <><div>Christopher Rasquin</div><header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1>Siddharth</h1>
-          <p>
-            Edit <code>src/App.tsx</code> and save to reload.
-          </p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-            Justin Burger
-          </a>
-        </header><Form>
-            <Form.Label>API Key:</Form.Label>
-            <Form.Control type="password" placeholder="Insert API Key Here" onChange={changeKey}></Form.Control>
-            <br></br>
-            <Button className="Submit-Button" onClick={handleSubmit}>
-              Submit
-            </Button>
-          </Form></>
-      }
-      <Footer page={page} setPage={setPage}></Footer>
-  </div>
+    <>
+      <Navbar page={page} setPage={setpage}></Navbar>
+      <Homepage page={page} setPage={setpage}></Homepage>
+      <Footer page={page} setPage={setpage}></Footer>
+    </>
   );
 }
